requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
aiohttp>=3.9.0
finnhub-python>=2.4.18

